# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETRA public safety platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  db?: string;
  ai?: string;
}

export interface Incident {
  id: string;
  zone: string;
  type: string;
  severity: number;
  threat_score: number;
  threat_behaviors?: string[];
  person_count: number;
  density: number;
  timestamp: string;
  resolved: boolean;
  /** @nullable */
  assigned_to?: string | null;
  sos_triggered: boolean;
  /** @nullable */
  status?: string | null;
}

export interface IncidentUpdate {
  resolved?: boolean;
  /** @nullable */
  assigned_to?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface IncidentListResponse {
  items: Incident[];
  total: number;
  page: number;
  pages: number;
}

export type IncidentStatsByType = {[key: string]: number};

export type IncidentStatsBySeverity = {[key: string]: number};

export type IncidentStatsByZone = {[key: string]: number};

export interface IncidentStats {
  total_today: number;
  active_alerts: number;
  zones_monitored: number;
  citizens_enrolled: number;
  by_type?: IncidentStatsByType;
  by_severity?: IncidentStatsBySeverity;
  by_zone?: IncidentStatsByZone;
}

export interface ConsentInput {
  citizen_name: string;
  phone: string;
  zone: string;
  language: string;
}

export interface ConsentRecord {
  id: string;
  citizen_name: string;
  phone: string;
  zone: string;
  language: string;
  active: boolean;
  created_at: string;
}

export interface SOSInput {
  /** @nullable */
  citizen_id?: string | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  voice_note_b64?: string | null;
  /** @nullable */
  photo_b64?: string | null;
}

export interface SOSReport {
  id: string;
  latitude: number;
  longitude: number;
  timestamp: string;
  resolved: boolean;
  reference: string;
}

export type ChatInputHistoryItem = { [key: string]: unknown };

export interface ChatInput {
  message: string;
  language: string;
  history?: ChatInputHistoryItem[];
}

export interface ChatResponse {
  reply: string;
  language_detected: string;
}

export interface ForecastPoint {
  time_label: string;
  predicted: number;
  upper: number;
  lower: number;
}

export interface AQIZone {
  zone: string;
  aqi: number;
  category: string;
}

export interface AQIResponse {
  zones: AQIZone[];
  co2_kg: number;
  events_count: number;
}

export type ListIncidentsParams = {
page?: number;
limit?: number;
zone?: string;
type?: string;
severity_min?: number;
severity_max?: number;
date_from?: string;
date_to?: string;
};

export type GetForecastParams = {
zone?: string;
};