# lib/api-zod/src/generated/types/aQIResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETRA public safety platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AQIZone } from './aQIZone';

export interface AQIResponse {
  zones: AQIZone[];
  co2_kg: number;
  events_count: number;
}