# lib/api-zod/src/generated/types/aQIZone.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETRA public safety platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AQIZone {
  zone: string;
  aqi: number;
  category: string;
}