# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETRA public safety platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatInputHistoryItem } from './chatInputHistoryItem';

export interface ChatInput {
  message: string;
  language: string;
  history?: ChatInputHistoryItem[];
}