# lib/api-zod/src/generated/types/consentRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETRA public safety platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsentRecord {
  id: string;
  citizen_name: string;
  phone: string;
  zone: string;
  language: string;
  active: boolean;
  created_at: string;
}