# lib/api-zod/src/generated/types/incidentListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETRA public safety platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Incident } from './incident';

export interface IncidentListResponse {
  items: Incident[];
  total: number;
  page: number;
  pages: number;
}