# lib/api-zod/src/generated/types/incidentStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETRA public safety platform API
 * OpenAPI spec version: 0.1.0
 */
import type { IncidentStatsBySeverity } from './incidentStatsBySeverity';
import type { IncidentStatsByType } from './incidentStatsByType';
import type { IncidentStatsByZone } from './incidentStatsByZone';

export interface IncidentStats {
  total_today: number;
  active_alerts: number;
  zones_monitored: number;
  citizens_enrolled: number;
  by_type?: IncidentStatsByType;
  by_severity?: IncidentStatsBySeverity;
  by_zone?: IncidentStatsByZone;
}