# lib/api-zod/src/generated/types/sOSInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETRA public safety platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SOSInput {
  /** @nullable */
  citizen_id?: string | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  voice_note_b64?: string | null;
  /** @nullable */
  photo_b64?: string | null;
}